// types/IViagens.ts
type IViagens = {
    id: string;
    id_cooperado: string;
    id_veiculo: string;
    data_marcacao: string;
    hora_marcacao: string;
    escolha: string;
    data_escolha: string;
    hora_escolha: string;
    estado_origem: string;
    cidade_origem: string;
    estado_destino: string;
    cidade_destino: string;
    distancia: string;
    obs: string;
    data_retorno: string;
    perde_vez: string;
    nome: string;
    descricao: string;
    tipo: string;
    placa: string;
    marca: string;
    modelo: string;
}

export default IViagens;